% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitening.r
\name{do_whitening}
\alias{do_whitening}
\title{Whiten the data using the specified method.}
\usage{
do_whitening(x, whiten)
}
\arguments{
\item{x}{A numeric matrix of data to be whitened.}

\item{whiten}{A character string specifying the whitening method to use.}
}
\description{
This function is used internally by the infomax function and is not intended for direct use by users.
}
\keyword{internal}
