% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax-package.R
\docType{package}
\name{infomax-package}
\alias{infomax}
\alias{infomax-package}
\title{infomax: Infomax Independent Component Analysis}
\description{
Implements Infomax and Extended Infomax Independent Component Analysis (ICA). A reimplementation of the mini-batch gradient descent algorithm found in 'EEGLAB' and 'MNE-Python', frequently used to perform ICA on electroencephalographic or magnetoencephalographic time-series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eegverse/infomax}
  \item Report bugs at \url{https://github.com/eegverse/infomax/issues}
}

}
\author{
\strong{Maintainer}: Matt Craddock \email{matt@mattcraddock.com}

Authors:
\itemize{
  \item Matt Craddock \email{matt@mattcraddock.com}
}

Other contributors:
\itemize{
  \item Scott Makeig et al., SCCN/INC/UCSD \email{scott@sccn.ucsd.ed} [copyright holder]
}

}
\keyword{internal}
