% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax-package.R
\docType{package}
\name{infomax-package}
\alias{infomax}
\alias{infomax-package}
\title{infomax: Infomax Independent Component Analysis}
\description{
Implements Infomax Independent Component Analysis (ICA)
}
\keyword{internal}
