% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax.R
\name{run_infomax}
\alias{run_infomax}
\title{Run Infomax ICA}
\usage{
run_infomax(
  x,
  centre = TRUE,
  pca = NULL,
  anneal = 0.98,
  annealdeg = 60,
  tol = 1e-07,
  lrate = NULL,
  blocksize = NULL,
  kurtsize = 6000,
  maxiter = 200,
  extended = TRUE,
  whiten = c("eeglab", "ZCA", "PCA", "ZCA-cor", "PCA-cor", "none"),
  verbose = TRUE
)
}
\arguments{
\item{x}{matrix of data; features in columns, samples in rows.}

\item{centre}{Mean-centre columns before running the algorithm.}

\item{pca}{Use PCA dimensionality reduction. Often helpful when the data is
rank deficient.}

\item{anneal}{Anneal rate at which learning rate reduced}

\item{annealdeg}{Angle at which learning rate reduced}

\item{tol}{Tolerance for convergence of ICA. Defaults to 1e-07.}

\item{lrate}{Initial learning rate.}

\item{blocksize}{size of blocks of data used for learning}

\item{kurtsize}{Size of blocks for kurtosis checking}

\item{maxiter}{Maximum number of iterations}

\item{extended}{Run extended-Infomax}

\item{whiten}{Whitening method to use}

\item{verbose}{print informative messages}
}
\value{
A list containing: * S  Matrix of source estimates * M  Estimated
mixing matrix * W  Estimated unmixing matrix * iter Number of iterations
completed
}
\description{
Run Infomax and extended-Infomax on a matrix of data. Mini-batch stochastic
gradient descent algorithm.
}
\references{
Bell, A.J., & Sejnowski, T.J. (1995). An information-maximization approach to blind separation and blind deconvolution. \emph{Neural Computation, 7,} 1129-159
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
