% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax.R
\name{run_infomax}
\alias{run_infomax}
\title{Run Infomax ICA}
\usage{
run_infomax(
  x,
  centre = TRUE,
  pca = NULL,
  anneal = 0.98,
  annealdeg = 60,
  tol = 1e-07,
  lrate = NULL,
  blocksize = NULL,
  kurtsize = 6000,
  maxiter = 200,
  extended = TRUE,
  whiten = c("ZCA", "PCA", "ZCA-cor", "PCA-cor", "none", "eeglab"),
  verbose = TRUE
)
}
\arguments{
\item{x}{matrix of data; features in columns, samples in rows.}

\item{centre}{mean-centre columns}

\item{pca}{Use PCA dimensionality reduction}

\item{anneal}{Anneal rate at which learning rate reduced}

\item{annealdeg}{Angle at which learning rate reduced}

\item{tol}{Tolerance for convergence of ICA}

\item{lrate}{Initial learning rate}

\item{blocksize}{size of blocks of data used for learning}

\item{kurtsize}{Size of blocks for kurtosis checking}

\item{maxiter}{Maximum number of iterations}

\item{extended}{Run extended-Infomax}

\item{whiten}{Whitening method to use}

\item{verbose}{print informative messages}
}
\description{
Run Infomax and extended-Infomax on a matrix of data. Mini-batch stochastic gradient descent algorithm.
}
