% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomax.R
\name{run_infomax}
\alias{run_infomax}
\title{Run Infomax ICA}
\usage{
run_infomax(
  x,
  centre = TRUE,
  pca = NULL,
  anneal = 0.98,
  annealdeg = 60,
  tol = 1e-07,
  lrate = NULL,
  blocksize = NULL,
  kurtsize = 6000,
  maxiter = 200,
  extended = TRUE,
  whiten = c("sqrtm", "ZCA", "PCA", "ZCA-cor", "PCA-cor", "none"),
  verbose = TRUE
)
}
\arguments{
\item{x}{matrix of data; features in columns, samples in rows.}

\item{centre}{Mean-centre columns before running the algorithm.}

\item{pca}{Use PCA dimensionality reduction. Often helpful when the data is
rank deficient.}

\item{anneal}{Anneal rate at which learning rate reduced}

\item{annealdeg}{Angle at which learning rate reduced}

\item{tol}{Tolerance for convergence of ICA. Defaults to 1e-07.}

\item{lrate}{Initial learning rate.}

\item{blocksize}{Size of blocks of data used for learning}

\item{kurtsize}{Size of blocks for kurtosis checking}

\item{maxiter}{Maximum number of iterations}

\item{extended}{Run extended-Infomax. Defaults to TRUE.}

\item{whiten}{Whitening method to use. See notes on usage.}

\item{verbose}{Print informative messages for each update of the algorithm.}
}
\value{
A list containing:
\itemize{
\item S  Matrix of source estimates
\item M  Estimated mixing matrix
\item W  Estimated unmixing matrix
\item iter Number of iterations completed
}
}
\description{
Run Infomax or extended-Infomax on a matrix of data. Mini-batch stochastic
gradient descent algorithm.
}
\references{
\itemize{
\item Bell, A.J., & Sejnowski, T.J. (1995). An information-maximization approach to blind separation and blind deconvolution. \emph{Neural Computation, 7,} 1129-159
\item Makeig, S., Bell, A.J., Jung, T-P and Sejnowski, T.J., "Independent component analysis of electroencephalographic data,"  In: D. Touretzky, M. Mozer and M. Hasselmo (Eds). Advances in Neural  Information Processing Systems 8:145-151, MIT Press, Cambridge, MA (1996).
}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
